refactor: split GraphQL ops, auth, comments, publications, MCP schemas, controller
Behavior-preserving hotspot splits: operation-definitions by scope, publication ingestion phases, comment mutations/target resolution barrels, auth route clusters, MCP tool-output schemas into domain modules, and WorkspacePageController action wiring extraction.
