+++ src/lib/mcp/tool-output-schemas/bus-schemas.ts
@@ -1,0 +1,7 @@
+import { z } from "zod";
+
+export const compactBusCampusSchema = z.strictObject({
+  id: z.number().int(),
+  namePrimary: z.string(),
+  nameSecondary: z.string().nullable(),
+});
